+++ src/intro.js
@@ -1,0 +1,286 @@
+import { START_PHONE_W, START_PHONE_H } from './ui.js';
+/* global truck, girl */
+import { lureNextWanderer, scheduleNextSpawn, queueLimit } from './entities/customerQueue.js';
+import { resumeWanderer } from './entities/wanderers.js';
+import { GameState } from './state.js';
+import { debugLog, DEBUG } from './debug.js';
+import { dur } from './ui.js';
+
+let startOverlay=null;
+let startButton=null;
+let phoneContainer=null;
+let startMsgTimers=[];
+let startMsgBubbles=[];
+
+function showStartScreen(scene){
+  scene = scene || this;
+  if (typeof debugLog === 'function') debugLog('showStartScreen called');
+  if(startButton){ startButton.destroy(); startButton = null; }
+  if(typeof phoneContainer !== 'undefined' && phoneContainer){
+    phoneContainer.destroy();
+    phoneContainer = null;
+  }
+  if(startOverlay){ startOverlay.destroy(); startOverlay = null; }
+  startMsgTimers.forEach(t => t.remove(false));
+  startMsgTimers = [];
+  startMsgBubbles.forEach(b => b.destroy());
+  startMsgBubbles = [];
+  startOverlay = scene.add.rectangle(240,320,480,640,0x000000,0.75)
+    .setDepth(14);
+
+  const phoneW = (typeof START_PHONE_W === 'number') ? START_PHONE_W : 260;
+  const phoneH = (typeof START_PHONE_H === 'number') ? START_PHONE_H : 500;
+  const caseG = scene.add.graphics();
+  caseG.fillStyle(0x5c3b2a,1);
+  caseG.fillRoundedRect(-phoneW/2-10,-phoneH/2-10,phoneW+20,phoneH+20,40);
+  const blackG = scene.add.graphics();
+  blackG.fillStyle(0x000000,1);
+  blackG.fillRoundedRect(-phoneW/2,-phoneH/2,phoneW,phoneH,30);
+  const whiteG = scene.add.graphics();
+  whiteG.fillStyle(0xffffff,1);
+  whiteG.fillRoundedRect(-phoneW/2+6,-phoneH/2+6,phoneW-12,phoneH-12,24);
+  const homeH = 100;
+  const homeG = scene.add.graphics();
+  homeG.fillStyle(0xf0f0f0,1);
+  homeG.fillRoundedRect(-phoneW/2+12,phoneH/2-homeH-12,phoneW-24,homeH,20);
+
+  const btnLabel = scene.add.text(0,0,'Clock In',{
+      font:'32px sans-serif',fill:'#fff'})
+    .setOrigin(0.5);
+  const bw = btnLabel.width + 60;
+  const bh = btnLabel.height + 20;
+  const btnBg = scene.add.graphics();
+  btnBg.fillStyle(0x007bff,1);
+  btnBg.fillRoundedRect(-bw/2,-bh/2,bw,bh,15);
+  const offsetY = phoneH/2 - homeH/2 - 12;
+  const containerY = 320;
+  phoneContainer = scene.add.container(240,containerY,[caseG,blackG,whiteG,homeG])
+    .setDepth(15);
+
+  const birdY = phoneH/2 - homeH - 60;
+  const bigBird1 = scene.add.sprite(-60,birdY,'sparrow2',0)
+    .setScale(3)
+    .setDepth(16);
+  const bigBird2 = scene.add.sprite(0,birdY,'sparrow2',0)
+    .setScale(3)
+    .setDepth(16);
+  const bigBird3 = scene.add.sprite(60,birdY,'sparrow2',0)
+    .setScale(3)
+    .setDepth(16);
+  bigBird1.anims.play('sparrow2_fly');
+  bigBird2.anims.play('sparrow2_ground');
+  bigBird3.anims.play('sparrow2_peck');
+  const bigBird4 = scene.add.sprite(-120,birdY,'sparrow3',0)
+    .setScale(3)
+    .setDepth(16);
+  const bigBird5 = scene.add.sprite(120,birdY,'sparrow3',0)
+    .setScale(3)
+    .setDepth(16);
+  bigBird4.anims.play('sparrow3_fly');
+  bigBird5.anims.play('sparrow3_ground');
+  phoneContainer.add([bigBird4,bigBird1,bigBird2,bigBird3,bigBird5]);
+
+  startButton = scene.add.container(0,offsetY,[btnBg,btnLabel])
+    .setSize(bw,bh)
+    .setInteractive({ useHandCursor: true });
+
+  const startZone = scene.add.zone(0,0,bw,bh).setOrigin(0.5);
+  startZone.setInteractive({ useHandCursor:true });
+  startButton.add(startZone);
+
+  phoneContainer.add(startButton);
+
+  let startMsgY = -phoneH/2 + 20;
+
+  const addStartMessage=(text)=>{
+    if(!phoneContainer) return;
+    const pad = 10;
+    const wrapWidth = phoneW - 60;
+    const txt = scene.add.text(0,0,text,{font:'20px sans-serif',fill:'#fff',wordWrap:{width:wrapWidth}})
+      .setOrigin(0,0.5);
+    const bw = txt.width + pad*2;
+    const bh = txt.height + pad*2;
+    const bg = scene.add.graphics();
+    bg.fillStyle(0x8bd48b,1);
+    bg.fillRoundedRect(-bw/2,-bh/2,bw,bh,10);
+    txt.setPosition(-bw/2 + pad, 0);
+    const xPos = -phoneW/2 + bw/2 + 20;
+    const yPos = startMsgY + bh/2;
+    const bubble = scene.add.container(xPos,yPos,[bg,txt]).setDepth(16).setAlpha(0);
+    phoneContainer.add(bubble);
+    startMsgBubbles.push(bubble);
+    startMsgY += bh + 10;
+    scene.tweens.add({targets:bubble,alpha:1,duration:300,ease:'Cubic.easeOut'});
+  };
+
+  if(scene.time && scene.time.delayedCall){
+    const msgOptions=[
+      ['u coming in? 🤔', 'where u at??', 'mornin ☀️'],
+      ['better not still be in bed 😜', 'yo coffee girl ☕', 'stop ghostin me'],
+      ['late night? 🥱💃', 'phone dead again? 🔋', 'omg wait till u hear about this guy 😏'],
+      ['u good?', 'hope everythin\'s chill', '…sry 😬']
+    ];
+    let delay=0;
+    for(const opts of msgOptions){
+      delay += Phaser.Math.Between(5000,15000);
+      const msg = Phaser.Utils.Array.GetRandom(opts);
+      startMsgTimers.push(scene.time.delayedCall(delay,()=>addStartMessage(msg),[],scene));
+    }
+  }
+  startZone.on('pointerdown',()=>{
+    if (typeof debugLog === 'function') debugLog('start button clicked');
+    startMsgTimers.forEach(t=>t.remove(false));
+    startMsgTimers=[];
+    startMsgBubbles=[];
+    const tl=scene.tweens.createTimeline({callbackScope:scene,onComplete:()=>{
+      if(startButton) startButton.destroy();
+      if(startOverlay){startOverlay.destroy(); startOverlay=null;}
+      phoneContainer.destroy(); phoneContainer=null;
+      playIntro.call(scene);
+    }});
+    tl.add({targets:phoneContainer,y:-320,duration:600,ease:'Sine.easeIn'});
+    tl.add({targets:startOverlay,alpha:0,duration:600},0);
+    tl.play();
+  });
+}
+
+function pauseWanderersForTruck(scene){
+  const threshold = 60;
+  GameState.wanderers.slice().forEach(c => {
+    if(!c.sprite) return;
+    if(Math.abs(c.sprite.x - truck.x) < threshold){
+      if(c.walkTween){
+        c.walkTween.stop();
+        c.walkTween.remove();
+        c.walkTween=null;
+      }
+      scene.tweens.add({targets:c.sprite,y:'-=20',duration:dur(150),yoyo:true});
+      scene.time.delayedCall(dur(1000),()=>{
+        if(GameState.girlReady && GameState.queue.length < queueLimit() && GameState.wanderers.includes(c)){
+          lureNextWanderer(scene, c);
+        }else{
+          resumeWanderer(scene, c);
+        }
+      },[],scene);
+    }
+  });
+}
+
+function playIntro(scene){
+  if(!truck || !girl) {
+    if (DEBUG) console.warn('playIntro skipped: missing truck or girl');
+    return;
+  }
+  if (typeof debugLog === 'function') debugLog('playIntro starting');
+  scene = scene || this;
+  if(!truck || !girl) return;
+  GameState.girlReady = false;
+  if(typeof debugLog==='function') debugLog('customers start spawning');
+  scheduleNextSpawn(scene);
+  const width = (scene.scale && scene.scale.width) ? scene.scale.width : 480;
+  const offscreenX = width + 100;
+  truck.setPosition(offscreenX,245).setScale(0.462);
+  girl.setPosition(offscreenX,245).setVisible(false);
+  const vibrateAmp = { value: 2 * (truck.scaleX / 0.924) };
+  const vibrateTween = (scene.tweens && scene.tweens.addCounter) ? scene.tweens.addCounter({
+    from: 0,
+    to: Math.PI * 2,
+    duration: dur(100),
+    repeat: -1,
+    onUpdate: t => {
+      const y = 245 + Math.sin(t.getValue()) * vibrateAmp.value;
+      if (truck.setY) {
+        truck.setY(y);
+      } else {
+        truck.y = y;
+      }
+    }
+  }) : { stop: ()=>{} };
+  if (scene.tweens && scene.tweens.add) {
+    scene.tweens.add({
+      targets: vibrateAmp,
+      value: 0,
+      duration: dur(600),
+      delay: dur(900),
+      ease: 'Sine.easeOut'
+    });
+  }
+  let smokeDelay = 100;
+  let smokeEvent = { remove: ()=>{} };
+  if (scene.time && scene.time.addEvent) {
+    smokeEvent = scene.time.addEvent({
+      delay: dur(smokeDelay),
+      loop: true,
+      callback: () => {
+        const puff = scene.add.text(truck.x + 60, truck.y + 20, '💨', { font: '20px sans-serif', fill: '#fff' })
+          .setDepth(1);
+        if (scene.tweens && scene.tweens.add) {
+          scene.tweens.add({
+            targets: puff,
+            x: puff.x + 30,
+            y: puff.y - 10,
+            alpha: 0,
+            duration: dur(800),
+            onComplete: () => puff.destroy()
+          });
+        } else {
+          puff.destroy();
+        }
+        smokeDelay += 125;
+        smokeEvent.delay = dur(smokeDelay);
+      }
+    });
+    scene.time.delayedCall(dur(1300), () => smokeEvent.remove(), [], scene);
+  }
+  const intro=scene.tweens.createTimeline({callbackScope:scene});
+  const hopOut=()=>{
+    const startX = truck.x + truck.displayWidth / 2 - 20;
+    const startY = truck.y - 10;
+    const endX = truck.x - 40;
+    const endY = 292;
+    const curve = new Phaser.Curves.QuadraticBezier(
+      new Phaser.Math.Vector2(startX, startY),
+      new Phaser.Math.Vector2(startX - 60, startY - 60),
+      new Phaser.Math.Vector2(endX, endY)
+    );
+    const follower={t:0,vec:new Phaser.Math.Vector2()};
+    scene.tweens.add({
+      targets:follower,
+      t:1,
+      duration:dur(700),
+      ease:'Sine.easeInOut',
+      onStart:()=>girl.setVisible(true),
+      onUpdate:()=>{
+        curve.getPoint(follower.t,follower.vec);
+        girl.setPosition(follower.vec.x,follower.vec.y);
+      },
+      onComplete:()=>{
+        girl.setPosition(endX,endY);
+        if(typeof debugLog==='function') debugLog('intro finished');
+        GameState.girlReady = true;
+        lureNextWanderer(scene);
+      }
+    });
+  };
+  intro.add({
+    targets:truck,
+    x:240,
+    scale:0.924,
+    duration:dur(1500),
+    ease:'Sine.easeOut',
+    onComplete:()=>{
+      smokeEvent.remove();
+      vibrateTween.stop();
+      if(truck.setY){
+        truck.setY(245);
+      }else{
+        truck.y=245;
+      }
+      pauseWanderersForTruck(scene);
+      hopOut();
+    }
+  });
+  intro.play();
+}
+
+export { showStartScreen, playIntro };
